+++ Socolin.RabbitMQ.Client.Tests.Integration/CancelActiveConsumerTests.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Newtonsoft.Json;
@@ -11,7 +9,7 @@ using Socolin.RabbitMQ.Client.Options.Consumer;
 
 namespace Socolin.RabbitMQ.Client.Tests.Integration
 {
-	public class LogExceptionTests
+	public class CancelActiveConsumerTests
 	{
 		private static readonly string BaseQueueName = $"Queue-{nameof(RabbitMqServiceClientTests)}";
 
@@ -54,34 +52,58 @@ namespace Socolin.RabbitMQ.Client.Tests.Integration
 		}
 
 		[Test]
-		public async Task CanUseDelayedConsumer()
+		public async Task CanCancelAndWaitCurrentTaskToBeCompleted()
 		{
-			var thrownException = new Exception();
-			var semaphore = new SemaphoreSlim(1);
-			var receivedExceptions = new List<(Exception exception, bool lastAttempt)>();
 			var consumerOptions = new ConsumerOptionsBuilder<string>()
 				.WithDefaultDeSerializer(message => JsonConvert.DeserializeObject<string>(Encoding.UTF8.GetString(message.Span)))
-				.WithFastRetryMessageAck(3)
-				.WithErrorLogging((exception, lastAttempt) => receivedExceptions.Add((exception, lastAttempt)))
+				.WithSimpleMessageAck()
 				.Build();
 
-			await _serviceClient.StartListeningQueueAsync(_queueName, consumerOptions, (message, _, ct) =>
+			var completed = false;
+			var messageCount = 0;
+			var activeConsumer = await _serviceClient.StartListeningQueueAsync(_queueName, consumerOptions, async (message, _, ct) =>
 			{
-				semaphore.Release();
-				return Task.FromException(thrownException);
+				messageCount++;
+				await Task.Delay(TimeSpan.FromSeconds(1), ct);
+				completed = true;
 			});
 
-			await semaphore.WaitAsync();
 			await _serviceClient.EnqueueMessageAsync(_queueName, "some-message");
-			await semaphore.WaitAsync();
-			await Task.Delay(TimeSpan.FromSeconds(1));
+			await _serviceClient.EnqueueMessageAsync(_queueName, "some-message");
+			await Task.Delay(TimeSpan.FromMilliseconds(200));
 
-			receivedExceptions.Should().BeEquivalentTo(
-				(thrownException, false),
-				(thrownException, false),
-				(thrownException, false),
-				(thrownException, true)
-			);
+			await activeConsumer.CancelAfterCurrentTaskCompletedAsync();
+
+			messageCount.Should().Be(1);
+			completed.Should().Be(true);
+		}
+
+
+		[Test]
+		public async Task CanCancelAndInterruptCurrentProcessing()
+		{
+			var consumerOptions = new ConsumerOptionsBuilder<string>()
+				.WithDefaultDeSerializer(message => JsonConvert.DeserializeObject<string>(Encoding.UTF8.GetString(message.Span)))
+				.WithSimpleMessageAck()
+				.Build();
+
+			var completed = false;
+			var messageCount = 0;
+			var activeConsumer = await _serviceClient.StartListeningQueueAsync(_queueName, consumerOptions, async (message, _, ct) =>
+			{
+				messageCount++;
+				await Task.Delay(TimeSpan.FromSeconds(1), ct);
+				completed = true;
+			});
+
+			await _serviceClient.EnqueueMessageAsync(_queueName, "some-message");
+			await _serviceClient.EnqueueMessageAsync(_queueName, "some-message");
+			await Task.Delay(TimeSpan.FromMilliseconds(200));
+
+			await activeConsumer.CancelAsync();
+
+			messageCount.Should().Be(1);
+			completed.Should().Be(false);
 		}
 	}
 }