+++ Socolin.RabbitMQ.Client/Exceptions/ProcessingAlreadyInProgressException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Socolin.RabbitMQ.Client.Exceptions
+{
+	public class ProcessingAlreadyInProgressException : Exception
+	{
+		public ProcessingAlreadyInProgressException(string message) : base(message)
+		{
+		}
+	}
+}