+++ Socolin.RabbitMQ.Client/Pipes/Consumer/Context/ConsumerPipeContext.cs
@@ -10,12 +10,19 @@ namespace Socolin.RabbitMQ.Client.Pipes.Consumer.Context
 		public T? DeserializedMessage { get; set; }
 		public BasicDeliverEventArgs RabbitMqMessage { get; }
 		public ProcessorMessageDelegate<T> MessageProcessor { get; }
+		public IActiveMessageProcessorCanceller ActiveMessageProcessorCanceller { get; }
 		public Dictionary<string, object> Items { get; } = new Dictionary<string, object>();
 
-		public ConsumerPipeContext(IModel chanel, BasicDeliverEventArgs basicDeliverEventArgs, ProcessorMessageDelegate<T> messageProcessor)
+		public ConsumerPipeContext(
+			IModel chanel,
+			BasicDeliverEventArgs basicDeliverEventArgs,
+			ProcessorMessageDelegate<T> messageProcessor,
+			IActiveMessageProcessorCanceller activeMessageProcessorCanceller
+		)
 		{
 			RabbitMqMessage = basicDeliverEventArgs;
 			MessageProcessor = messageProcessor;
+			ActiveMessageProcessorCanceller = activeMessageProcessorCanceller;
 			Chanel = chanel;
 		}
 	}