+++ Socolin.RabbitMQ.Client/Pipes/Consumer/Context/IConsumerPipeContext.cs
@@ -1,17 +1,19 @@
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 
 namespace Socolin.RabbitMQ.Client.Pipes.Consumer.Context
 {
-	public delegate Task ProcessorMessageDelegate<in T>(T message, Dictionary<string, object> items) where T : class;
+	public delegate Task ProcessorMessageDelegate<in T>(T message, Dictionary<string, object> items, CancellationToken cancellationToken) where T : class;
 
 	public interface IConsumerPipeContext<T> where T : class
 	{
 		IModel Chanel { get; }
 		BasicDeliverEventArgs RabbitMqMessage { get; }
 		T? DeserializedMessage { get; set; }
+		IActiveMessageProcessorCanceller ActiveMessageProcessorCanceller { get; }
 
 		/// <summary>
 		/// A function called at the end of the pipe to process the deserialized message